requirements.txt
absl-py==0.7.1
astor==0.8.0
blessings==1.7
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.17.0
nvidia-ml-py3==7.352.0
protobuf==3.9.1
psutil==5.6.3
PyYAML==5.1.2
scikit-learn==0.21.3
scipy==1.3.0
six==1.12.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
Werkzeug==0.15.5
wrapt==1.11.2
prettytable==0.7.2
ipdb==0.13.9